doctrineのアノテーション定義の読み込みと競合するため、EntityEventアノテーションをsrc/Annotationに移動
