+++ src/Eccube/Annotation/PostLoad.php
@@ -22,15 +22,15 @@
  */
 
 
-namespace Eccube\Entity\Event\Annotations;
+namespace Eccube\Annotation;
 
 
 use Doctrine\ORM\Events;
 
 /**
  * @Annotation
- * @TargetEvent(Events::preRemove)
+ * @TargetEvent(Events::postLoad)
  */
-final class PreRemove extends EntityEvent
+final class PostLoad extends EntityEvent
 {
 }
