+++ src/Eccube/Annotation/PostPersist.php
@@ -22,7 +22,7 @@
  */
 
 
-namespace Eccube\Entity\Event\Annotations;
+namespace Eccube\Annotation;
 
 
 use Doctrine\ORM\Events;
