+++ tests/Eccube/Tests/Entity/Event/EntityEventDispatcherTest.php
@@ -5,9 +5,9 @@ namespace Eccube\Tests\Entity\Event;
 use Doctrine\ORM\Event\LifecycleEventArgs;
 use Doctrine\ORM\Events;
 use Eccube\Entity\Customer;
-use Eccube\Entity\Event\Annotations\PrePersist;
-use Eccube\Entity\Event\Annotations\PreUpdate;
-use Eccube\Entity\Event\Annotations\TargetEntity;
+use Eccube\Annotation\PrePersist;
+use Eccube\Annotation\PreUpdate;
+use Eccube\Annotation\TargetEntity;
 use Eccube\Entity\Event\EntityEventDispatcher;
 use Eccube\Entity\Event\EntityEventListener;
 use Eccube\Entity\Product;
